fix: remap TimescaleDB to port 5433 and fix ROUND cast in views
- Remap host port 5432 -> 5433 to avoid conflict with local PostgreSQL
- Fix ROUND(double precision, int) -> ROUND(::numeric, int) in v_active_service_gaps
- Update .env, .env.example, docker-compose.yml, and README accordingly

+++ sql/migrations/003_grafana_views.sql
@@ -97,7 +97,7 @@ SELECT
     r.route_type,
     sga.direction_id,
     ROUND(sga.gap_duration_sec::numeric, 0) AS gap_duration_sec,
-    ROUND(sga.gap_duration_sec / 60.0, 1)   AS gap_duration_min
+    ROUND((sga.gap_duration_sec / 60.0)::numeric, 1)   AS gap_duration_min
 FROM transit.service_gap_alerts sga
 LEFT JOIN transit.routes r ON r.route_id = sga.route_id
 WHERE sga.gap_start > NOW() - INTERVAL '60 minutes'
